Refactor FXIOS-13262 [Merino] Updating Merino provider logic with default number of stories to cache (#28873)
* change merino to fetch a larger number of stories

* fix existing tests and add new ones

* add a test

* forgot a print() in there lol

* capitalize L

+++ firefox-ios/Providers/Merino/MerinoFeedFetcher.swift
@@ -7,7 +7,7 @@ import MozillaAppServices
 
 protocol MerinoFeedFetching: Sendable {
     func fetch(
-        itemCount: Int32,
+        itemCount: Int,
         locale: CuratedRecommendationLocale,
         userAgent: String
     ) async -> [RecommendationDataItem]
@@ -18,7 +18,7 @@ struct MerinoFeedFetcher: MerinoFeedFetching {
     let logger: Logger
 
     func fetch(
-        itemCount: Int32,
+        itemCount: Int,
         locale: CuratedRecommendationLocale,
         userAgent: String
     ) async -> [RecommendationDataItem] {
@@ -29,7 +29,7 @@ struct MerinoFeedFetcher: MerinoFeedFetching {
                     userAgentHeader: userAgent
                 )
             )
-            let request = CuratedRecommendationsRequest(locale: locale, count: itemCount)
+            let request = CuratedRecommendationsRequest(locale: locale, count: Int32(itemCount))
             let response = try client.getCuratedRecommendations(request: request)
             return response.data
         } catch let error as CuratedRecommendationsApiError {
